整体优化代码结构

+++ server/like-admin/src/main/java/com/mdd/admin/vo/article/ArticleCateVo.java
@@ -14,7 +14,7 @@ public class ArticleCateVo implements Serializable {
 
     private Integer id;        // 主键
     private String name;       // 分类名称
-    private Integer number;    // 文章数量
+    private Long number;    // 文章数量
     private Integer sort;      // 排序编号
     private Integer isShow;    // 是否显示
     private String createTime; // 创建时间
